refactor(mountains): extract rock scenery + collision into mountains/rocks.ts
Move all rock code — the RockPosition type, the collision constants and
rockCollisionRadius / rockIsCollisionHazard rules, the shared craggy normal map,
applyRockSnowColors, the ROCK_STONES palette + makeRockColor, createRock, and
addRocks — into src/mountains/rocks.ts. It reads the analytic terrain samplers from
terrain.js; createTerrain (still in the hub for now) imports addRocks from here.

The hub re-exports RockPosition so the public ./mountains.js surface (used by
scene-setup.ts) is unchanged. Verified byte-identical: full Node suite green,
tree-collision 3/3, physics-invariant harness OK, verify 49/49, tsc + lint clean.
